split fciqmc.jl into multiple files; use user installed julia executable if available
